% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fun.R
\name{get_fun}
\alias{get_fun}
\title{Find a function}
\usage{
get_fun(fn, default = identity)
}
\arguments{
\item{fn}{A function name, either a character string or an unquoted
function name, with or without \link[base:ns-dblcolon]{colons}.}

\item{default}{If \code{fn} is \code{\link{NULL}}, the \code{default} function is returned.
Defaults to \code{\link[=identity]{identity()}}.}
}
\value{
A function
}
\description{
Find a function
}
\examples{
get_fun(plu_ral)
get_fun(plu::ral)
get_fun("plu_ral")
get_fun("plu::ral")

get_fun(NULL)
get_fun(NULL, default = plu_ral)
}
