% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plu-package.R
\docType{package}
\name{plu-package}
\alias{plu}
\alias{plu-package}
\title{plu: Pluralize Phrases}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Converts English phrases to singular or plural form
    based on the length of an associated vector.  Contains helper
    functions to create natural language lists from vectors and to include
    the length of a vector in natural language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/plu}
  \item Report bugs at \url{https://github.com/rossellhayes/plu/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} [copyright holder]

Other contributors:
\itemize{
  \item Kevin Atkinson \email{kevina@gnu.org} [data contributor]
}

}
\keyword{internal}
