% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plu_stick.R
\name{plu_stick}
\alias{plu_stick}
\alias{stick}
\title{Deprecated functions}
\usage{
plu_stick(...)

stick(...)
}
\arguments{
\item{...}{\lifecycle{deprecated}}
}
\value{
A \link[lifecycle:deprecate_soft]{deprecation error}.
}
\description{
\lifecycle{deprecated} This function has been deprecated in favor of
\code{\link[and:and]{and::and()}}, \code{\link[knitr:combine_words]{knitr::combine_words()}} or \code{\link[glue:glue_collapse]{glue::glue_collapse()}}.
}
